requirements.txt
flask
openai
requests
python-dotenv
pymongo
isodate
gradio